AAP-87798 refactor: move backfill_object_ansible_id to its own module
repair.py is for data corruption fixes; backfill.py is for population
of newly-added denormalized fields.
